% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Loads a CSV file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to the CSV file (character)}
}
\value{
The function returns a tibble (data.frame) based on the CSV file.
}
\description{
The function loads a CSV file defined by \code{filename} argument and returns
a tibble. If the path is incorrect the function will end with an error.
}
\examples{
\dontrun{
accident_2015 <- fars_read("./data/accident_2015.csv.bz2")
}

}

