% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Counts number of accidents per month and year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector or list of years (numeric or integer) that will be
searched in the data}
}
\value{
Returns a pivot tibble (data frame) with months in rows and selected
years in columns containing the number of accidents. Returns a warning for
every input year that does not exist in the datasets. Returns an error (no
results) if a different than numeric or integer input is presented.
}
\description{
Based on the list of years, the function calculates the number of accidents
in the US on a monthly basis. The accident files need to be in the working
directory, the years can be passed as a list or a vector.
}
\examples{
\dontrun{
fars_summarize_years(2013:2016)
}

}

