% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Numerical of integer input indicating a year of a data set.}
}
\value{
Returns a character string in a format "accident_<year>.csv.bz2" that
can be used as a file name
}
\description{
The function creates a filename for a .csv.bz2 file based on the \code{year}
argument in a form "accident_<year>.csv.bz2". It requires a numerical or
integer input otherwise ends with an error.
}
\examples{
\dontrun{
makefilename(2016)
}
}

